Spike/SpeakerDiarizationPython/requirements.txt
# PyAnnote speaker diarization
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0

# Audio processing
soundfile>=0.12.0
librosa>=0.10.0

# Utilities
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
